App/requirements.txt
# Dependencias del Aplicativo Visualización de Tablas CGA
# Instalar con: pip install -r requirements.txt

# Framework web
Flask>=3.0.0

# Base de datos MySQL
Flask-MySQLdb>=2.0.0
mysqlclient>=2.2.0

# Autenticación y sesiones
Flask-Login>=0.6.0
Werkzeug>=3.0.0

# Variables de entorno
python-dotenv>=1.0.0

# Servidor WSGI para producción (recomendado)
waitress>=3.0.0
